Move all legacy tests into a single location
